requirements.txt
Jinja2==3.1.4
PyYAML==6.0.2
python-slugify==8.0.4
requests==2.32.3
